requirements.txt
discord.py==1.2.3
Pillow==6.1.0
aiohttp==3.5.4